+++ Runtime/NumberFormat.cs
@@ -1,0 +1,11 @@
+namespace UnderLogic.Serialization.Toml
+{
+    public enum NumberFormat
+    {
+        Decimal,
+        HexLowerCase,
+        HexUpperCase,
+        Octal,
+        Binary
+    }
+}